Finish formatting form and input fields
I used flex-direction and wrap to get the input fields and their labels
in the right spots. I still need to swap the names around however to
match the example given by TOP. I also added text to the div above the
form and created the button and div w/text underneath the form.