Add password validation and style button
I styled the button and added an offset border when the user clicks On
the button. I used javascript to validate the input in the password and
confirm password fields when the Create Account button is clicked. If
the passwords don't match, an error is shown saying that the passwords
don't match. In addition, the border of the  field turns red. The field
turns back to its original color when the correct password is input.